feat(account): implement add-wallet command and move to common utils
